File Upload
1

+++ src/main/java/com/winter/app/regions/RegionDTO.java
@@ -6,12 +6,12 @@ public class RegionDTO {
 	//3. 생성자는 여어개 가능하지만, 기본생성자는 필수
 	//4. 멤버변수의 이름은 Datatype DB의 Table의 컬럼명과 Datatype 일치 
 	
-	private int region_id;
+	private Long region_id;
 	private String region_name;
-	public int getRegion_id() {
+	public Long getRegion_id() {
 		return region_id;
 	}
-	public void setRegion_id(int region_id) {
+	public void setRegion_id(Long region_id) {
 		this.region_id = region_id;
 	}
 	public String getRegion_name() {
